requirements.txt
# click==8.1.3
# Flask==2.2.3
# gunicorn==20.1.0
click==8.1.3
Flask==2.0.1
gunicorn==20.1.0
Werkzeug==2.2.2

# pandas==2.1.0
# numpy==1.26.0
# scikit-learn==1.5.2
# biopython==1.84
pandas
numpy
scikit-learn
biopython